[*] Fixed sign comparsion

+++ core/test/dap_common_test.c
@@ -882,9 +882,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_mul(*(signed char*)l_a, *(int*)l_b) == dap_mul_builtin(*(signed char*)l_a, *(int*)l_b), "MUL SIGNED CHAR and INT");
     dap_assert_PIF(dap_mul(*(unsigned char*)l_a, *(int*)l_b) == dap_mul_builtin(*(unsigned char*)l_a, *(int*)l_b), "MUL UNSIGNED CHAR and INT");
     dap_assert_PIF(dap_mul(*(unsigned short*)l_a, *(int*)l_b) == dap_mul_builtin(*(unsigned short*)l_a, *(int*)l_b), "MUL UNSIGNED SHORT and INT");
-    dap_assert_PIF(dap_mul(*(unsigned int*)l_a, *(int*)l_b) == dap_mul_builtin(*(unsigned int*)l_a, *(int*)l_b), "MUL UNSIGNED INT and INT");
-    dap_assert_PIF(dap_mul(*(unsigned long*)l_a, *(int*)l_b) == dap_mul_builtin(*(unsigned long*)l_a, *(int*)l_b), "MUL UNSIGNED LONG and INT");
-    dap_assert_PIF(dap_mul(*(unsigned long long*)l_a, *(int*)l_b) == dap_mul_builtin(*(unsigned long long*)l_a, *(int*)l_b), "MUL UNSIGNED LONG LONG and INT");
+    dap_assert_PIF((unsigned int)dap_mul(*(unsigned int*)l_a, *(int*)l_b) == (unsigned int)dap_mul_builtin(*(unsigned int*)l_a, *(int*)l_b), "MUL UNSIGNED INT and INT");
+    dap_assert_PIF((unsigned long)dap_mul(*(unsigned long*)l_a, *(int*)l_b) == (unsigned long)dap_mul_builtin(*(unsigned long*)l_a, *(int*)l_b), "MUL UNSIGNED LONG and INT");
+    dap_assert_PIF((unsigned long long)dap_mul(*(unsigned long long*)l_a, *(int*)l_b) == (unsigned long long)dap_mul_builtin(*(unsigned long long*)l_a, *(int*)l_b), "MUL UNSIGNED LONG LONG and INT");
     // long
     dap_assert_PIF(dap_mul(*(char*)l_a, *(long*)l_b) == dap_mul_builtin(*(char*)l_a, *(long*)l_b), "MUL CHAR and LONG");
     dap_assert_PIF(dap_mul(*(short*)l_a, *(long*)l_b) == dap_mul_builtin(*(short*)l_a, *(long*)l_b), "MUL SHORT and LONG");
@@ -894,9 +894,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_mul(*(signed char*)l_a, *(long*)l_b) == dap_mul_builtin(*(signed char*)l_a, *(long*)l_b), "MUL SIGNED CHAR and LONG");
     dap_assert_PIF(dap_mul(*(unsigned char*)l_a, *(long*)l_b) == dap_mul_builtin(*(unsigned char*)l_a, *(long*)l_b), "MUL UNSIGNED CHAR and LONG");
     dap_assert_PIF(dap_mul(*(unsigned short*)l_a, *(long*)l_b) == dap_mul_builtin(*(unsigned short*)l_a, *(long*)l_b), "MUL UNSIGNED SHORT and LONG");
-    dap_assert_PIF(dap_mul(*(unsigned int*)l_a, *(long*)l_b) == dap_mul_builtin(*(unsigned int*)l_a, *(long*)l_b), "MUL UNSIGNED INT and LONG");
-    dap_assert_PIF(dap_mul(*(unsigned long*)l_a, *(long*)l_b) == dap_mul_builtin(*(unsigned long*)l_a, *(long*)l_b), "MUL UNSIGNED LONG and LONG");
-    dap_assert_PIF(dap_mul(*(unsigned long long*)l_a, *(long*)l_b) == dap_mul_builtin(*(unsigned long long*)l_a, *(long*)l_b), "MUL UNSIGNED LONG LONG and LONG");
+    dap_assert_PIF((unsigned int)dap_mul(*(unsigned int*)l_a, *(long*)l_b) == (unsigned int)dap_mul_builtin(*(unsigned int*)l_a, *(long*)l_b), "MUL UNSIGNED INT and LONG");
+    dap_assert_PIF((unsigned long)dap_mul(*(unsigned long*)l_a, *(long*)l_b) == (unsigned long)dap_mul_builtin(*(unsigned long*)l_a, *(long*)l_b), "MUL UNSIGNED LONG and LONG");
+    dap_assert_PIF((unsigned long long)dap_mul(*(unsigned long long*)l_a, *(long*)l_b) == (unsigned long long)dap_mul_builtin(*(unsigned long long*)l_a, *(long*)l_b), "MUL UNSIGNED LONG LONG and LONG");
     // long long
     dap_assert_PIF(dap_mul(*(char*)l_a, *(long long*)l_b) == dap_mul_builtin(*(char*)l_a, *(long long*)l_b), "MUL CHAR and LONG LONG");
     dap_assert_PIF(dap_mul(*(short*)l_a, *(long long*)l_b) == dap_mul_builtin(*(short*)l_a, *(long long*)l_b), "MUL SHORT and LONG LONG");
@@ -906,9 +906,9 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_mul(*(signed char*)l_a, *(long long*)l_b) == dap_mul_builtin(*(signed char*)l_a, *(long long*)l_b), "MUL SIGNED CHAR and LONG LONG");
     dap_assert_PIF(dap_mul(*(unsigned char*)l_a, *(long long*)l_b) == dap_mul_builtin(*(unsigned char*)l_a, *(long long*)l_b), "MUL UNSIGNED CHAR and LONG LONG");
     dap_assert_PIF(dap_mul(*(unsigned short*)l_a, *(long long*)l_b) == dap_mul_builtin(*(unsigned short*)l_a, *(long long*)l_b), "MUL UNSIGNED SHORT and LONG LONG");
-    dap_assert_PIF(dap_mul(*(unsigned int*)l_a, *(long long*)l_b) == dap_mul_builtin(*(unsigned int*)l_a, *(long long*)l_b), "MUL UNSIGNED INT and LONG LONG");
-    dap_assert_PIF(dap_mul(*(unsigned long*)l_a, *(long long*)l_b) == dap_mul_builtin(*(unsigned long*)l_a, *(long long*)l_b), "MUL UNSIGNED LONG and LONG LONG");
-    dap_assert_PIF(dap_mul(*(unsigned long long*)l_a, *(long long*)l_b) == dap_mul_builtin(*(unsigned long long*)l_a, *(long long*)l_b), "MUL UNSIGNED LONG LONG and LONG LONG");
+    dap_assert_PIF((unsigned int)dap_mul(*(unsigned int*)l_a, *(long long*)l_b) == (unsigned int)dap_mul_builtin(*(unsigned int*)l_a, *(long long*)l_b), "MUL UNSIGNED INT and LONG LONG");
+    dap_assert_PIF((unsigned long)dap_mul(*(unsigned long*)l_a, *(long long*)l_b) == (unsigned long)dap_mul_builtin(*(unsigned long*)l_a, *(long long*)l_b), "MUL UNSIGNED LONG and LONG LONG");
+    dap_assert_PIF((unsigned long long)dap_mul(*(unsigned long long*)l_a, *(long long*)l_b) == (unsigned long long)dap_mul_builtin(*(unsigned long long*)l_a, *(long long*)l_b), "MUL UNSIGNED LONG LONG and LONG LONG");
     // signed char
     dap_assert_PIF(dap_mul(*(char*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(char*)l_a, *(signed char*)l_b), "MUL CHAR and SIGNED CHAR");
     dap_assert_PIF(dap_mul(*(short*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(short*)l_a, *(signed char*)l_b), "MUL SHORT and SIGNED CHAR");
@@ -918,7 +918,7 @@ static void s_test_overflow_diff_types(unsigned long long *l_a, unsigned long lo
     dap_assert_PIF(dap_mul(*(signed char*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(signed char*)l_a, *(signed char*)l_b), "MUL SIGNED CHAR and SIGNED CHAR");
     dap_assert_PIF(dap_mul(*(unsigned char*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(unsigned char*)l_a, *(signed char*)l_b), "MUL UNSIGNED CHAR and SIGNED CHAR");
     dap_assert_PIF(dap_mul(*(unsigned short*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(unsigned short*)l_a, *(signed char*)l_b), "MUL UNSIGNED SHORT and SIGNED CHAR");
-    dap_assert_PIF(dap_mul(*(unsigned int*)l_a, *(signed char*)l_b) == dap_mul_builtin(*(unsigned int*)l_a, *(signed char*)l_b), "MUL UNSIGNED INT and SIGNED CHAR");
+    dap_assert_PIF((unsigned int)dap_mul(*(unsigned int*)l_a, *(signed char*)l_b) == (unsigned int)dap_mul_builtin(*(unsigned int*)l_a, *(signed char*)l_b), "MUL UNSIGNED INT and SIGNED CHAR");
     dap_assert_PIF((unsigned long)dap_mul(*(unsigned long*)l_a, *(signed char*)l_b) == (unsigned long)dap_mul_builtin(*(unsigned long*)l_a, *(signed char*)l_b), "MUL UNSIGNED LONG and CHAR");
     dap_assert_PIF((unsigned long long)dap_mul(*(unsigned long long*)l_a, *(signed char*)l_b) == (unsigned long long)dap_mul_builtin(*(unsigned long long*)l_a, *(signed char*)l_b), "MUL UNSIGNED LONG LONG and SIGNED CHAR");
     // unsigned char
